split and refactored Pagescraper and scraper.ts
