+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetActivity.ts
