+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetClassId.ts
