+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetMode.ts
