+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetNeedsConsent.ts
