+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetSection.ts
