+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetStatus.ts
