+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassHelpers/GetTerm.ts
