+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/TimeHelpers/GetDays.ts
