+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/TimeHelpers/GetInstructor.ts
