+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/TimeHelpers/GetTime.ts
